requirements.txt
fastapi
loguru
requests
uvicorn
python-dotenv
